requirements.txt

requests==2.28.1
web3==5.28.0
python-telegram-bot==13.12
